split files to create a proper structure and update build scripts
